fixed typo in seed.rb

+++ db/migrate/20150422020117_create_locations.rb
@@ -4,7 +4,7 @@ class CreateLocations < ActiveRecord::Migration
       t.string :name
       t.float :latitude
       t.float :longitude
-      t.float :postcode
+      t.integer :postcode
     end
   end
 end
